+++ api/V0ScaleServer.go
@@ -64,7 +64,6 @@ func V0ScaleK3S(w http.ResponseWriter, r *http.Request) {
 
 	logrus.Debug("HTTP GET V0ScaleK3S: ", string(d))
 
-	w.WriteHeader(http.StatusAccepted)
 	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
 	w.Header().Set("Api-Service", "v0")