+++ init.go
@@ -38,7 +38,7 @@ func init() {
 	config.Credentials.Password = os.Getenv("AUTH_PASSWORD")
 	config.AppName = "Mesos K3S Framework"
 	config.K3SCustomDomain = os.Getenv("K3S_CUSTOM_DOMAIN")
-	config.K3SServerString = util.Getenv("K3S_SERVER_STRING", "--cluster-cidr \"10.1.0.0/16\" --service-cidr \"10.2.0.0/16\" --cluster-dns \"10.2.0.10\" --disable=traefik,servicelb,metrics-server --snapshotter=native --flannel-backend=host-gw --flannel-iface=ethwe ")
+	config.K3SServerString = util.Getenv("K3S_SERVER_STRING", "--cluster-cidr \"10.1.0.0/16\" --service-cidr \"10.2.0.0/16\" --cluster-dns \"10.2.0.10\" --disable=metrics-server --snapshotter=native --flannel-backend=ipsec --flannel-iface=ethwe ")
 	config.K3SAgentString = util.Getenv("K3S_AGENT_STRING", "--snapshotter=native --flannel-backend=host-gw --flannel-iface=ethwe ")
 	config.ImageK3S = util.Getenv("IMAGE_K3S", "docker.io/rancher/k3s:v1.20.0-k3s2")
 	config.ImageETCD = util.Getenv("IMAGE_ETCD", "docker.io/bitnami/etcd:latest")